+++ src/components/Breadcrumbs/Breadcrumb/interfaces.ts
@@ -1,30 +1,5 @@
-export interface IBreadcrumb {
-  colors: IBreadcrumbColors;
-}
-
-export interface IBreadcrumbColors {
-  primary: string;
-  secondary: string;
-  tertiary?: string;
-}
-
-export interface IBreadcrumbColorProps {
-  doneColor?: string;
-  undoneColor?: string;
-  titleColor?: string;
-}
-
-export interface IBreadcrumbBarProps extends IBreadcrumbColorProps {
-  active?: boolean;
-  done: boolean;
-}
-
-export interface IBreadcrumbProps extends IBreadcrumbBarProps {
-  hasNext?: boolean;
-  hasPrevious?: boolean;
+export interface IBreadcrumbProps {
   text: string;
-}
-
-export interface IBreadcrumbLineProps extends IBreadcrumbBarProps {
-  visible?: boolean;
+  isLast: boolean;
+  onClick?: () => void;
 }