+++ src/components/Breadcrumbs/index.stories.tsx
@@ -2,17 +2,13 @@ import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/react/types-6-0';
 
-import Breadcrumbs from './';
+import Breadcrumbs from '.';
 import { IBreadcrumbsProps } from './interfaces';
 
 export default {
   title: 'Breadcrumbs',
   component: Breadcrumbs,
-  argTypes: {
-    doneColor: { control: 'color' },
-    undoneColor: { control: 'color' },
-    titleColor: { control: 'color' },
-  },
+  argTypes: {},
 } as Meta;
 
 const Template: Story<IBreadcrumbsProps> = (args) => <Breadcrumbs {...args} />;
@@ -20,8 +16,9 @@ const Template: Story<IBreadcrumbsProps> = (args) => <Breadcrumbs {...args} />;
 export const Default = Template.bind({});
 Default.args = {
   breadcrumbs: [
-    { active: false, text: 'STEP ONE' },
-    { active: true, text: 'STEP TWO' },
-    { active: false, text: 'STEP THREE' },
+    { text: 'First page', onClick: () => alert("First page clicked") },
+    { text: 'Second page', onClick: () => alert("Second page clicked") },
+    { text: 'Third page', onClick: () => alert("Third page clicked") },
+    { text: 'Current page' },
   ],
 };