+++ src/components/Icons/index.ts
@@ -25,4 +25,8 @@ export { default as Remove } from './Remove';
 export { default as Plus } from './Plus';
 export { default as Download } from './Download';
 export { default as Delete } from './Delete';
+export { default as Share } from './Share';
+export { default as Alert } from './Alert';
+export { default as MagnifyingGlass } from './MagnifyingGlass';
+export { default as Chain } from './Chain';
 export { default as Information } from './Information';