+++ src/components/TagSelect/interfaces.ts
@@ -1,0 +1,12 @@
+export type tagOption = {
+  id: string;
+  text: string;
+  color: string;
+  background: string;
+};
+
+export interface ITagSelectProps {
+  options?: tagOption[];
+  selected: tagOption;
+  onChange?: (o: tagOption) => void;
+}