+++ src/components/Tooltip/index.stories.tsx
@@ -1,6 +1,5 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/react/types-6-0';
-import { Fragment } from 'react';
 import Tooltip from './';
 import { ITooltipProps } from './interfaces';
 import PublicProfileIcon from '../Icons/PublicProfile';